backend/requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile /app/requirements/base.in
#
anyio==3.6.2
    # via starlette
asgiref==3.6.0
    # via django
attrs==22.2.0
    # via pytest
build==0.10.0
    # via pip-tools
click==8.1.3
    # via
    #   pip-tools
    #   strawberry-graphql
    #   uvicorn
django==4.1.6
    # via
    #   -r /app/requirements/base.in
    #   django-debug-toolbar
    #   strawberry-graphql-django
django-debug-toolbar==3.8.1
    # via -r /app/requirements/base.in
exceptiongroup==1.1.0
    # via pytest
factory-boy==3.2.1
    # via -r /app/requirements/base.in
faker==16.7.0
    # via factory-boy
graphql-core==3.2.3
    # via strawberry-graphql
h11==0.14.0
    # via uvicorn
idna==3.4
    # via anyio
iniconfig==2.0.0
    # via pytest
libcst==0.4.9
    # via strawberry-graphql
markdown-it-py==2.1.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
mypy-extensions==1.0.0
    # via typing-inspect
packaging==23.0
    # via
    #   build
    #   pytest
pillow==9.4.0
    # via -r /app/requirements/base.in
pip-tools==6.12.2
    # via -r /app/requirements/base.in
pluggy==1.0.0
    # via pytest
psycopg2==2.9.5
    # via -r /app/requirements/base.in
pygments==2.14.0
    # via
    #   rich
    #   strawberry-graphql
pyproject-hooks==1.0.0
    # via build
pytest==7.2.1
    # via
    #   -r /app/requirements/base.in
    #   pytest-django
pytest-django==4.5.2
    # via -r /app/requirements/base.in
python-dateutil==2.8.2
    # via
    #   faker
    #   strawberry-graphql
python-multipart==0.0.5
    # via strawberry-graphql
pyyaml==6.0
    # via libcst
rich==13.3.1
    # via
    #   -r /app/requirements/base.in
    #   strawberry-graphql
six==1.16.0
    # via
    #   python-dateutil
    #   python-multipart
sniffio==1.3.0
    # via anyio
sqlparse==0.4.3
    # via
    #   django
    #   django-debug-toolbar
starlette==0.24.0
    # via strawberry-graphql
strawberry-graphql[debug-server]==0.155.4
    # via
    #   -r /app/requirements/base.in
    #   strawberry-graphql-django
strawberry-graphql-django==0.9.2
    # via -r /app/requirements/base.in
tomli==2.0.1
    # via
    #   build
    #   pyproject-hooks
    #   pytest
typing-extensions==4.4.0
    # via
    #   libcst
    #   strawberry-graphql
    #   typing-inspect
typing-inspect==0.8.0
    # via libcst
uvicorn==0.20.0
    # via strawberry-graphql
wheel==0.38.4
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
